Tighten the error SLO for the git service
Today there was an incident caused by a software defect when rolling
out a feature flag. The engineer rolling out the changes to Gitaly was
watching Gitaly metrics, but not clientside metrics. This caused them
to miss the problem.

If we had tighter alerts for the git service at the time, we would
have caught this sooner and perhaps not have had to wait for a manual
report by a QA run.