fix: Use instance as join label for redis-ops rates
Before we were using `pod` and `fqdn` as the join labels. This would
mean that when a `pod` had the same name across environments, we'd
select series for both environments and display them on the graph.

Changing that to use the `instance` label, which uses an
`ip-adress:port` in the k8s case solves this.