fix(alert): ignore sentry for chef_client_disabled
What
---
Ignore the `type=sentry` for the `ChefClientDisabled` alert.

Why
---
`chef-client` was disabled on sentry in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/7761. We are
in the process of migrating to the new setup in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15372. In
the mean time we should ignore this.

When https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15372
is done we can remove this negation, but it wouldn't be a big deal if we
keep it either because Sentry will be deployed on Kubernetes now.

Reference: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/7946
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>