Allow missing feature_category label if selector still works
We were forbidding a recording rule lookup when the selector contained
labels that don't match the recording rule, which makes sense. But in
the case of application SLIs, we do have an edge case: the
`feature_category` selector maybe a regex that allows the empty
string (which matches a missing label).

This allows that specific case: if the only missing label is
`feature_category`, and the selector contains a `feature_category`
selector that matches the empty string, we'll use that recording rule.
This means that the `global_search` SLI can use its recordings on the
dashboards.