fix: improve alertmanager alerts recovery time
What
---
Update the interval from 1h to 5m

Why
---
When this alert fires it can take up to 1hour to resolve since we are
using the 1h window. Instead, use a 5m window which will give us faster
recovery times.

It might end up getting a bit noisy, but we can improve that in a
future merge request to make it less sensitive.

Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>