Allow filtering SLIs on dashboards by feature category
Some application SLIs have multiple feature categories available from
source metrics, but not _all_ feature categories are possible values.
For instance, for customersdot, only feature categories from the
Fulfillment stage are possible.

Having customersdot metrics show on other group's dashboards is a bit
noisy, so this change allows an SLI to define
`dashboardFeatureCategories` and filter by those when deciding to put it
on a dashboard or not.