Adjust post-deploy migration metric
Post-deploy migrations are executed independently from the auto-deploy
pipeline, for that reason it's important for release managers to know if
there are pending post-deploy migrations to be executed. The pressure of
post-migrations can be visible on the release manager dashboard
https://dashboards.gitlab.net/d/delivery-release_management/delivery-release-management?orgId=1&refresh=5m.

However this number is not accurate because it considers both databases
(`main` and `ci`) when calculating the pending migrations, effectively
duplicating the actual number of pending post-deploy migrations.

This commit fixes that situation by splitting the total by 2.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2498