Remove sentry metric from release manager dashboard
The sentry metric reports the number of sentry issues per environment,
while informative these numbers are not actinoable from the release
manager side. Furthermore, release managers and the production health
system in release-tools don't rely on sentry issues to decide if a
deployment should be promoted or not.

This commit removes this metric in order to make this dashboard more
actionable for the release managers and the Delivery team.