fix: ignore ListBranchNamesContainingCommit ratelimit
What
---
Ignore the `Unavailable` grpc code returned from the `ListBranchNamesContainingCommit` RPC.

Why
---
In https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16844 we
are seeing a single user being rate limited which is by design. We are
considereing `Unavailable` as an error when it shouldn't be since we are
rate limiting by design.

In https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16844 we
plan to update the error code returned by Gitaly to `ResourceExhausted`
but in the mean time ignore it not to page the on-call.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16844
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>