Fix similar problem with output from toolbox
The toolbox invocation implicitly combines STDERR and STDOUT.
That means parse errors from stackcollapse (which are common and tolerable)
get included in the flamegraph SVG output file (which breaks it due to the
XML becoming invalid).

To fix, we move the file generation to also happen inside the toolbox
container, leaving only STDERR to go to toolbox's terminal output.