Order the saturation metrics
This gives us a fixed ordering in the table below the saturation
graph. Ideally the ordering would be:

- Hard SLO
- Soft SLO
- P99 Quantile
- P95 Quantile
- All series ordered by their highest peak in the shown range.

This makes it easier to see in the table which series pused the
quantile higher.

Seeing which series breached SLO also becomes easier, as those will be
ordered above the SLOs in the table.