fix: clamp occurence SLA ratio between 1 and 0
On 2023-01-09 we had some observability issues that caused some
recordings to not complete. As a result, this means that the
success-side (numerator) of the fraction could be higher than
operation-side (denominator). This caused availability to be higher
than 100%.

The opposite could also happen when we miss recordings for the
ops-rate for error SLIs.

This doesn't solve incorrect recordings, but it does make numbers less
ridiculous when they do happen, causing less confusion if they only
happen for a brief period of time. I don't think this affects
instances collect all metrics in a single Prometheus instance.