chore: update capacity planning SLO for redis-cpu to 85%
At the time of writing, this would mean that we'd have a capacity
planning issue for `redis`, `redis-cache`, `redis-ratelimiting`.

This means that we'll automatically close the capacity planning issue
for `redis-sessions`.

For `redis-cache`, the issue might be closed after the recent sharding
or `repository-cache`.

For `redis-ratelimiting` the issue might be closed if the traffic
pattern settles after the seasonal dip we had at the end of the year.

Current issues:

- `redis`: https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/545
- `redis-cache`: https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/364
- `redis-ratelimiting`: https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/562
- `redis-sessions`: https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/576