Record crossover mappings for removed stage groups
This generates a new file,
`services/stage-group-mapping-crossover.jsonnet`, every time we remove
feature categories. The old categories are the keys of an object. If a
category was renamed, we can set the value to the new category name, and
we'll treat those two names as the same thing for the stage group
purposes. For example, if we have:

    {
      api: 'rest_and_graphql_api',
    }

Then whichever stage group has `rest_and_graphql_api` will also count
`api` towards its totals, while we shake out those changes from the
Rails app. The new time we update this repo, the file can go back to
being an empty object (`{}`).