feat: add environment to use for capacity planning
This adds the environment to use for capacity planning to the
saturation manifest.

We will then be able to consume that in Tamland, which currently only
gathers saturation points from `gprd`. But with `thanos` now running
in the `thanos` environment, we need to include that environment for
some saturation points.

For example:

```
(master → origin {20} U:3 ?:1 ✗) runbooks scripts/compile_jsonnet.rb metrics-catalog/saturation/tamland.jsonnet | jq .shard_cpu.capacityPlanningEnvironments
[
  "gprd",
  "thanos"
]
```