feat: Add recording rule duration SLI
This adds a new SLI to both Thanos and Monitoring (for
Prometheus). This SLI keeps an eye on rule-group durations: every time
a duration exceeds it's interval will be counted as an error, meaning
we need to work on improving the duration of that rule group.

We're recording this as an apdex, because this SLI talks about the
latency (duration) of a recording rule group.