fix/thanos-compactor-down-runbook
What
-----------------------------------------------------------------
Updates runbook for troubleshooting thanos-compactor down issue
due to overlapping blocks, to remove a script for determining target
blocks for deletion.

Why
-----------------------------------------------------------------
Noticed during an [incident](https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8413)
that the script pulls up 1st block from the array of overlapping blocks
as a target of deletion but it is not always the case to target 1st block.