doc(gitaly): how to export a git repository
What
---
Create a new runbook on how to create a git repository export and share
it with developers so that they can debug any problems they are facing
with the repository.

Why
---
We've got asked multiple times by the Gitaly team to get a git
repository export so that they can debug a problem with a git
repository. Doing is not documented and is triabl knowledge. In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16833 we
plan to make this self-serve/automated but we can't automate what we
don't have documented.

Examples:
- https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16828
- https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16674
- https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/17469

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/17469
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>