Fix saturation point for shared runners
Before this, we were measuring `shared`-runner utilization as follows:

```
gitlab_runner_jobs
/
gitlab_runner_concurrent
```

While the `private` and `shared-gitlab-org` shards we were doing:

```
gitlab_runner_jobs
/
gitlab_runner_limit
```

Because the labelset for `gitlab_runner_concurrent` does not include a
`system_id` label, we were not recording for
`gitlab:component_saturation:ratio` for the `shared` shard.

With this, we're using the same metric for the denominator for all
shards (`gitlab_runner_limit`).

We're also making sure that we're including all shards into she
`shared_runners` saturation component by only excluding `private` and
`shared-gitlab-org` shards from that component. This means we'll start
including the `saas-linux` shards and the macOS and windows
runners. Because of this, we needed to add the `shard` label as a
significant label, so we can distinguish which shard we're predicting
saturation for in dashboards.

This has all been merged into a single file to keep the saturation
point definitions the same for all of these shards. The only thing
that changed for `private_runners` and `shared_runners_gitlab` is the
wording in the description.