chore: move all apdex metrics into their own directory
This collects all of the apdex-type metrics into their own
directory. We intend to add another one, which will make it a third
one, and extract some of the shared parts into a library.

In this we also rename `rateApdex` to `successCounterApdex`. Allowing
us to add an `errorCounterApdex` next to it.