Exclude gitaly_total_disk_pace from capacity planning
We have better signals for disk space in Gitaly:

- `gitaly_active_node_available_space`: Will warn us if the set of
gitaly nodes receiving new projects is running out of disk space.
- `disk_space`: Will warn us if any disk (including Gitaly ones) is
running out of space. So we'll cover any other Gitaly node that is not
actively receiving traffic separately.

The total disk space utilization is interesting to record in terms of
efficiency: we don't want to have too much free space as to not pay to
much for unused storage, but we don't need to alert or capacity plan
for it.