feat: separate thanos recordings by environment
This allows specifying a `separateGlobalRecordingSelectors` on the
metrics-config object using a `{ <name>: <selectorHash> }` mapping. By
doing that, we'll create a recording rule group in a separate file for
each of the specified selectors.

Each selector will be merged into the selectors specified for the
source-aggregation set for that specific rule group.

Separating these will make the separate rule-groups faster, as a slow
Prometheus in a non-prod environment will not slow down a rule
execution for production. This will also allow us to place
`partial_response_strategy: abort` on some rule groups that require
accurate data, as the abort would not mean missing metrics when a
non-prod Prometheus fails to reply.