refactor: file per global recording environment
This adds a helper function to generate a file per environment for
recording rules in thanos.

It takes a function as an argument. The function should receive a
selector hash, and should return a hash with in this format:

   basename: "yamlcontent generated using the passed selector"

This will then be transformed into a hash like this:

   basename-ops.yml: "yamlcontent generated using the passed selector"
   basename-gprd.yml: "yamlcontent generated using the passed selector"
   basename-other.yml: "yamlcontent generated using the passed selector"

This removes the need for us to repeat how the filename is built, and
we'd only have to change it in one place if we decided to do that.