feat(dashboard): show gitaly CPU throttling
What
---
Add a new time-series panel to show cgroups that are being CPU throttled.
If the data is empty then there is no throttling.

Why
---
With the rollout of `cfs_quota_us` per repository we need to make it
easy for the operator that a cgroup is getting throttled.

We only show the cgroup that is getting throttled and when it's getting
throttled.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/17332
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>