fix(gitaly): reduce apdex SLO from 0.999 to 0.97
What
---
Set the apdex from 99.9% to 97%

Why
---
Looking at the [last month
pages/incident](https://gitlab.com/gitlab-com/gl-infra/reliability-reports/-/issues/166)
for Gitaly it was one of the busiest services. Most of these incidents
resolved on their own and were ~severity::3 or ~severity::4.

This is adding a lot of burden to our on-call and burning them out, we
have https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/991 to
address some of the stability issues.

In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23576#note_1379544308
we can see how setting the apdex to `0.97` wouldn't have paged the
on-call for a ~severity::4 incident.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23576
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>