Use min(slo_observation_statu) for per service availability
This changes the way we display per-service availability in the SLA
dashboard to match how we calculate the overall system availability.

The overall system availability is recorded using

    min without(slo_type) (slo_observation_status)

While on the SLA dashboard we were showing an average of both the
`apdex` and `error` slo type. This caused the service-availability
numbers to look higher than what we would use for our overall SLA
calculation.