Customize weight on the occurence SLA dashboard
This adds templates (dropdowns) to the occurence SLA dashboard for
each service in the SLA calculation. The templates are prefilled with
the values stored in the service catalog but can be adjusted to
anything we want.

All service recordings need to be present for this dashboard to be
able to render values.