Add SLA panel with customizable weights
This adds an extra panel and templates to the general SLA dashboard.

The templates allow selecting a weight per service, and the small SLA
panel will display the overall system availability with the weights
selected in the templates.

There might be a small drift between the values from the source
metric (`slo_observation_status`) and the recorded SLA. These are
rounding errors that should not exceed 0.01%, but they could cause the
panels to colour differently.