Don't use recordingRuleRegistry for some metrics
Not all Prometheus instances, like the one in ai-assist are in our
cluster. As a result, they don't have all our recording rules
deployed.

This also means that they can't use metrics from the recording rule
registry (`sli_aggregations:`).

To bypass this, this commit allows bypassing the registry by
specifying that when adding the SLI.

In this we also make sure that we don't adjust the key metric
recordings that could be in another service's key-metrics recording
file based on the significant labels for these SLIs.