Evaluate code_suggestions SLIs in Thanos
Since the Prometheus instance does not get the recording rules
deployed. We can do the evaluation in Thanos for now.

We'll have to revisit this once we've moved the cluster and are able
to have part of the rules evaluated in Prometheus.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23722