Add `env` label to promSource aggregation
This adds the `env` label to the promSource aggregation. We select for
this on dashboards and we use this aggregation set for the services
evaluated in Thanos.

This shouldn't matter for the services evaluated in Prometheus,
because those already have the `env` label anyway because of the
prometheus instance they come from.

Another piece in
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23722