Extend code_suggestions.model_gateway SLI
This renames the `server` SLI to `model_gateway`

It also drops the severity to S4 while in beta.

We're also adding an apdex for the `model_gateway` SLI with a
`satisfiedThreshold` of 10s. Anything using the 1s bucket and we
wouldn't meet SLO when if we exclude the 4xx responses. Which tend to
be the fastest.

It also replaces the request counter (`http_requests_total`) with the
`http_request_duration_seconds_count` summary.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23725