Fix uploading dashboards
The problem occurred when trying to upload the code_suggestions-triton
dashboard. Which accidentally matched another dashboard we had. The
update would then fail because we'd be trying to update a dashboard
with in mismatched folder.

This removes all hardcoded ids from our dashboards and ensures we
don't accidentally add them again.