Labeled kube-state-metrics for globally evaluated services
This kube-state-metrics recording rules for globally evaluated
services. This will add `:labeled` version of that that contain our
known `tier` & `type` labels.

For things like code-suggestions (running in the `ai-assist` cluster)
this will be better than relying on the global `type` label that is
also on the metrics, because it will allow us to separate monitoring
infrastructure (`prometheus`) from the application for saturation
metrics. Currently the the metrics from prometheus pods running in
`ai-assist` all have the `type="code_suggestions"` label, in the
`:labeled` variant this will be overwritten with `type="monitoring"`.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23724