Include env in thanos evaluated saturation points
Before this change, we would no longer have the env label, which is on
the source metrics in the resulting saturation point. We do use that
label as a filter in other places, for example the dashboards so make
sure it is also there.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23724