Lower the severity of gcp_quota_limit for code suggestions
While the service is not generally available, we're labelling all
alerts as S4 to no notify the on-call engineers. Sadly, the saturation
recordings and alerts are aggregated across services, which means we
can't easily set a separate severity on them.

This deals with the problem by creating a new saturation point for
just the `s4` portion.