Select only env=ops for thanos saturation
This adjusts the way the saturation point queries are built to not
ignore selectors passed that are also static labels. The static labels
could be overriding labels that we do want to select for in saturation
points. In that case, the selector passed in should still apply.

The same is applicable for the promql query we build for alerts. Those
get used to generate images using Grafana and need the filter with the
original selector, not the overridden static labels.

As discussed in
https://gitlab.com/gitlab-com/gl-infra/capacity-planning/-/issues/1051#note_1440566238