Stage group aggregation for Thanos evaluated services
This adds the `gitlab:component:stage_group` and `gitlab:stage_group`
aggregations for globally evaluated services.

In the parent commit, we've added the feature category metrics for
those. They do not have a `monitor="default"` label like they would
have in the case where the evaluation happened it Prometheus.

But we also don't want to add all of the
`gitlab:component:feature_category:*` metrics that are already being
aggregated from the prometheus source ones, so we add a type selector
to only include Thanos evaluated services.

This would be much simpler if we did all evaluation globally in
thanos as described in https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2303#proposal