fix(alerts): remove LoggingVisibilityDiminished
What
---
Delete the `LoggingVisibilityDiminished` alert.

Why
---
Looking at the last 5 incidents:

- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/15969
- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/15942
- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/15825
- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/14543
- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/14407

For all of them, the EOC just had to wait and do no action, sometimes even
silencing the alert. At the moment this alert is
[unactionable](https://gitlab.com/gitlab-com/runbooks/-/blob/95deba25ee75020960a5929fbc835d06c1f82ba6/docs/monitoring/deleting-alerts.md#L18-25)
and has no customer-facing impact. The only impact is engineers at
GitLab Inc. seeing delayed logs, but the on-call doesn't do anything.

There is
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16480 to
investigate the issue and see if we can implement something. We can
always add this after if we fix the scaling issues.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/16480
Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>