Update metric recording rules for SaaS runners
We're in the middle of renaming `shared` fleet to
`saas-linux-small-amd64` to match naming of our other runner shards.
Switch from old shard to new one is done within
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/15991.

Our recording rules are however hardcoded to use `shared`. That means
not only they will stop working once we will switch the traffic, but
they are also not tracking our other publicly offered shards that we
offer.

This commit will update the metrics to use new shard naming and to add
missing shards to the metrics catalog.