Exclude workhorse from service aggregation
This excludes the workhorse component from service aggregations in
favour of the rails request component.

The rails request component has more appropriate apdex targets defined
inside the application itself. Which makes it more appropriate to use
here.

The disadvantage of this approach is that any 503 errors that
workhorse might encounter because a puma worker is not available would
not be recognized as errors for the service. I prefer this tradeoff,
because when we would not have enough workers available like this,
this would also be noticeable in the apdex. For GitLab.com this is
unlikely because we have 1:1 workhorse-puma in the same pod.

This only affects the service aggregation that rolls up into the SLA
for GitLab.com, it does not affect alerting.