Avoid duplicate sidekiq sli_aggregation recordings
The metrics for the recording rule registry, that reduces cardinality
for use in SLIs are already recorded in the main
`rules/autogenerated-key-metrics-sidekiq.yml` file. These intermediate
recording rules are not scoped to a `type`, which means that they are
being recorded twice.

This removes that duplicate recording in the `ops-gitlab-net` file,
and makes the sidekiq SLIs for ops-gitlab-net rely on the other
recordings.