Draft: Introduce component-level information for Tamland
We'd like to be able to add information to the service and the component
level for Tamland. A component here is a service's concrete saturation
point, e.g. `patroni + pg_primary_cpu`.

I'd like to see if this makes sense to add to the service catalog
directly or if we should find a place "on the sides" for this (e.g. in
tamland.jsonnet).