Use a 1h range in the apdex-ratio experiment
The 1m range isn't used in SLA calculations anymore. For that we use
the 5m range.

The 1h range is used the most for error budgets for stage groups and
for the mwmbr alerts. Theoretically, this should be a worse case in
terms of data to load, but most of this data should be cached minute
to minute.

This also matches the range already in the file for the ratelimiting
experiment.

From the comment in: https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6111/diffs#note_1502377404