Add experimental recording rules with offset
This adds a version of the recording rules with a 1m offset as an
experiment to see if we can work around the jitter we're seeing in
apdexes. The 1m offset here is rather unsatisfying because it adds
another 1m of delay in recording rules, but this one 1m would at least
be controlled rather than stacked up from recording rules.

More info in https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2445#note_1503906053