use abort for partial responses in experiment
This uses `partial_response_strategy: abort` for the experimental
rules. I don't think this will make a lot of difference, because my
understanding is that the partial responses strategy only affects the
recordings when an entire thanos-sidecar fails to respond, which I
don't think was the case in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2445#note_1503294764

But I do think we should try to get to a place where our default is
`partial_response_strategy: abort` rather than `warn` if we move to a
world where all the recording happens in thanos. This is why we
separate recordings per environment now.