Specify offset on source metrics aggregation set
This allows adding an `offset:` attribute to an aggregation set that
collects source metrics.

The offset is specified in the form of a PromQL duration, and the
validation ensures that it is.

When an offset is specified, the recording rules generated for the
aggregation set will include that specified duration offsetting the
source metrics used for the recording.

This merge request applies this for the `featureCategory` and the
`promSourceSLIs` aggregation sets that are currently evaluated in
prometheus. This does add some delay to the metrics that are visible,
but we'll reduce this again when we move the metrics to Thanos in
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/1098.

This only works for aggregating metrics directly from source,
aggregation sets that are used to transform other metrics don't
implement this yet.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2482