Replace pg_int4_id with new metrics
This saturation metric uses new metrics introduced in
https://gitlab.com/gitlab-cookbooks/gitlab-exporters/-/merge_requests/309/. This
new metric includes all integer columns in the database, rather than
only the selected ones that the old version saturation metric
did.

At the moment of writing, we'd be at ~51%, utilization
because of `system_notes_metadata.id`. The same as the old saturation
point. This is being worked on in
https://gitlab.com/gitlab-org/gitlab/-/issues/424114

Iid columns are being ignored as discussed in
https://gitlab.com/gitlab-org/gitlab/-/issues/421329

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/421329