Record transactional rates in aggregation set transformations
A transformed aggregegation set is an aggregation set that is not
sourced directly from source metrics, but instead is sourced from
another aggregation set.

This allows recording transactional rates in aggregation set
transformations. When specifying `errorRates` or `apdexRates` on a
transformed aggregation set. When the source aggregation also has
`errorRates` or `apdexRates` specified.

When this is specified, the recording rules will include a simple
recording rule that adds the `recorded_rate` label on top of the other
specified aggregation labels.

This does not yet use these transactional rates to record the
ratio. Which is also a requirement of
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2483 but
will be done in a follow up as to not grow this MR even more.