Globally evaluate rules for services using Runway
These services are using a Prometheus to which we don't deploy any
recording rules. Instead Prometheus uses thanos-receive to get all of
the metrics scraped from the running service into Thanos.

This means that we need to evaluate and deploy the recording rules
there.

This applies this to the Runway-archetype, and as a result we modify
the recording rule definitions for the `ai-gateway` service.