The ai-gateway uses old openmetrics histograms
The `code_suggestions` one was using `10.0` but the `ai-gateway` is
using `10`.

This is likely due to different Prometheus client versions. The
`code_suggestions` service is using the newer version, so it is likely
to come back in the future. That is why we're setting the state to
"migrating" for the histograms in the `ai-gateway` service.