fix: Adjust Runway latency metrics
The underlying metric behaves like a gauge, instead of a counter (see
discussion at
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/89#note_1546764061).
To accomodate this, allow overwriting the `rangeVectorFunction` used in
our histograms per Apdex definition.

We also adjust to the actual values of the latency buckets in GCP's
metric by adjusting our thresholds to existing values, and changing the
graph y axis to milliseconds instead of seconds.