Add llm-completion SLI to Sidekiq
This adds the llm-completion SLI to the sidekiq service. This SLI
should eventually move to the AI gateway, but as the AI gateway is
currently not yet involved in these operations, this adds it to the
Sidekiq service that emits the metrics.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2507