Record transactional rates from source metrics
This allows specifying `errorRates` or `apdexRates` on an
aggregation-set that will be used for aggregating from source metrics.

Doing this will record both sides of what will become a ratio in a
single recording rule with a `recorded_rate` label added using
`label_replace` signifying which rate is being recorded.

The recording rules would look as following for error rates:

```yaml
record: 'source_error:rates_5m',
expr: |
  label_replace(
    sum by (a,b) (
      rate(some_error_total_count{}[5m] offset 2s)
    )
    or
    (
      0 * sum by (a,b) (
        rate(some_total_count{}[5m] offset 2s)
      )
    ),
    'recorded_rate', 'error_rate' , '', ''
  )
  or
  label_replace(
    sum by (a,b) (
      rate(some_total_count{}[5m] offset 2s)
    ),
    'recorded_rate', 'ops_rate' , '', ''
  )
```

Note the fallback to `0 * ops_rate` in the `error_rate` portion. This
is done to make sure that we record an error rate of 0, even when the
error rate is missing as long as the operation rate is present. We do
this because more often than not or error rates are not properly
initialized. If we didn't have this fallback, we wouldn't be able to
record an error ratio until we've seen an error.

The recording rule for apdex ratios looks as follows:

```yaml
record: 'source_apdex:rates_5m',
expr: |
  label_replace(
    sum by (a,b) (
      rate(some_apdex_success_total_count{}[5m] offset 2s)
    ),
    'recorded_rate', 'success_rate' , '', ''
  )
  or
  label_replace(
    sum by (a,b) (
      rate(some_apdex_total_count{}[5m] offset 2s)
    ),
    'recorded_rate', 'apdex_weight' , '', ''
  )
```

Here we aren't using the fallback because we expect most of our apdex
measurements to be successful, meaning the metric would not often be
missing.