Avoid passing aggregationLabels with aggregationSet
We were often passing both the aggregagationLabels around as an
argument next to the aggregation set. The aggregationLabels are a
property of the aggregationSet already, so it was confusing to have
both in functions without knowing where they came from.

As we can see, this changes nothing in the existing recording rules.

Noticed this when working on
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6243