Allow generating rule files in nested directories
This allows generating rules in subdirectories for different
environments.

This is not yet supported by our current Thanos and prometheus setup,
that receive their rules through the syncinator. But the new thanos
setup, already in use for `thanos-staging`, does.

It also updates the `make generate` script to delete rule files in
subdirectories so they don't get left behind when regenerating.