[Thanos-staging] Transactional rates from source metrics
This adds experimental recording rules for recording aggregation-sets
from source metrics in thanos staging.

The experimental aggregation set includes the new transactional rates
besides the old rates we currently record in Prometheus. This is done
for 4 services to start.

This does not yet allow recordin ratio's from the source metrics or in
transformations. This will be done in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2475 later
when we get rid of the intermediate recording rules in Prometheus.