Add a thanos-staging service
This adds a new service `thanos-staging` to the metrics- and
service-catalog.

This will make it so we generate the same SLI rules for
`thanos-staging` as we do for `thanos`. However, we cannot view these
metrics in Grafana, as they are only present in `thanos-staging` and
not in the production `thanos` that backs Grafana.

To do this, we've added a `thanosType` field to
`dangerouslyThanosEvaluated` services. When the `thanosType` is
`thanos-staging`, we'll only generate and evaluate the SLI rules for
the service in staging. For the other globally evaluated services,
we're using `thanosType='thanos'`. This field will likely be
refactored when we move all recording rules to thanos in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2495.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/24423